Feat: color set Asset에 추가
