+++ MadiiWidget/Resources/Color/Color+madiiColor.swift
